feat: all commands inside json
