+++ index.js
@@ -1,6 +1,6 @@
 import { InputListener } from './src/servises/InputListener.js';
 import { StateManager } from './src/servises/StateManager.js';
-import { CommandValidator } from './src/servises/CommandValidator.js';
+import { CommandIdentifier } from './src/servises/CommandIdentifier.js';
 import { CommandExecutor } from './src/servises/CommandExecutor.js';
 // TODO make import dom index
 
@@ -9,7 +9,7 @@ import { FileManager } from './src/FileManager.js';
 const fileManager = new FileManager(
   InputListener,
   StateManager,
-  CommandValidator,
+  CommandIdentifier,
   CommandExecutor,
 );
