+++ src/servises/CommandExecutor.js
@@ -1,10 +1,10 @@
 import { CommandProvider } from './CommandProvider.js';
-import { CommandValidator } from './CommandValidator.js';
+import { default as availableCommands } from '../commands.json' assert { type: 'json' };
 
 export class CommandExecutor extends CommandProvider {
   constructor() {
     super();
-    this.commands = CommandValidator.commands.reduce((availableCommands, { input, method }) => {
+    this.commands = availableCommands.reduce((availableCommands, { input, method }) => {
       availableCommands[input] = this[method];
       return availableCommands;
     }, {});
